app/requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==20.3.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==8.1.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
kombu==5.0.2
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.3
pytest-django==4.2.0
pytest-factoryboy==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
